fix bug: customer delete

+++ app/Http/Controllers/CustomerController.php
@@ -124,12 +124,16 @@ class CustomerController extends Controller
     public function destroy($id)
     {
         $user = User::find($id);
-        
-        foreach ($user->affiliate_user->coupon as $coupon) {
-            # code...
-            $coupon->delete();
+        // dd($user->affiliate_user->coupon);
+
+        if($user->affiliate_user) {
+            foreach ($user->affiliate_user->coupon as $coupon) {
+                # code...
+                $coupon->delete();
+            }
+            $user->affiliate_user->delete();
         }
-        $user->affiliate_user->delete();
+
         $user->delete();
         flash(translate('Customer has been deleted successfully'))->success();
         return redirect()->route('customers.index');
